requirements.txt
django
django-debug-toolbar
djangorestframework
psycopg2-binary
